nanglerng/requirements.txt
fontawesomefree==6.0.0
django
django-quill-editor
django-taggit